// Frontend/src/app/models/reporte.model.ts
export interface ReporteClientePorcino {
  clienteCedula: string;
  clienteNombre: string;
  clienteApellido: string;
  clienteTelefono: string;
  clienteDireccion: string;
  porcinoIdentificacion: string;
  porcinoEdad: number;
  porcinoPeso: number;
  razaNombre: string;
  alimentacionDescripcion: string;
  alimentacionDosis: number;
}